luce_vm/luce_django/luce/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# luce/accounts/models.py: 7
# luce/lucehome/serializers.py: 2
# luce/lucehome/urls.py: 3,4
# luce/lucehome/views.py: 6,8,9,15,17,18
djangorestframework == 3.12.4
